Finished QFlag class for KLM/POD and moved methods _get_corrupt_mask and get_qual_flags to the parent class Reader
