+++ pygac/reader.py
@@ -656,7 +656,20 @@ class Reader(six.with_metaclass(ABCMeta)):
         QFlag = self.QFlag
         if flags is None:
             flags = QFlag.FATAL_FLAG | QFlag.CALIBRATION | QFlag.NO_EARTH_LOCATION
-        return (self.scans['quality_indicators'] & flags.value).astype(bool)
+        return (self.scans['quality_indicators'] & int(flags)).astype(bool)
+
+    def get_qual_flags(self):
+        """Read quality flags."""
+        number_of_scans = self.scans["telemetry"].shape[0]
+        qual_flags = np.zeros((int(number_of_scans), 7))
+        qual_flags[:, 0] = self.scans["scan_line_number"]
+        qual_flags[:, 1] = self._get_corrupt_mask(flags=self.Qflag.FATAL_FLAG)
+        qual_flags[:, 2] = self._get_corrupt_mask(flags=self.Qflag.CALIBRATION)
+        qual_flags[:, 3] = self._get_corrupt_mask(flags=self.Qflag.NO_EARTH_LOCATION)
+        qual_flags[:, 4] = self._get_corrupt_mask(flags=self.Qflag.CH_3_CONTAMINATION)
+        qual_flags[:, 5] = self._get_corrupt_mask(flags=self.Qflag.CH_4_CONTAMINATION)
+        qual_flags[:, 6] = self._get_corrupt_mask(flags=self.Qflag.CH_5_CONTAMINATION)
+        return qual_flags
 
     @abstractmethod
     def postproc(self, channels):
